highbd_variance64: use unsigned math w/sse calculation
This makes the operation defined should out of range input be used in
e.g., fuzzing tests. In practice, the library expects 8/10/12-bit values
to be in range ([0, 255], [0, 1023], [0, 4095]). This allows data types
to be chosen that facilitate SIMD optimizations.

Bug: 475394379

+++ vpx_dsp/variance.c
@@ -250,7 +250,7 @@ static void highbd_variance64(const uint8_t *src8_ptr, int src_stride,
     for (j = 0; j < w; ++j) {
       const int diff = src_ptr[j] - ref_ptr[j];
       *sum += diff;
-      *sse += diff * diff;
+      *sse += (uint32_t)diff * diff;
     }
     src_ptr += src_stride;
     ref_ptr += ref_stride;
